feat : create real initial migration
With the new migration system, during development developers are forced to generate a migration file.
The easiest way to do that is to use the autogenerate function of alembic. However this needs to have a database initialised with tables from before any changement was done compared to main.

Therefor, developers are currently forced to switch to main to use the auto-creation function (they can't use their branch as it would build the database as they want it after the migration) and then switch back to their branch to use the autogenerate.
This can seems simple however when working on branch with updated dependencies bringing breaking changes, it implies dependencies have to be uninstalled when switching to main to reinstall the old one and then use the inversed operation to switch back on the branch.

With this file, the database can be built without using the auto-creation system by using the function `alembic upgrade head`