Topic notification transition (#412)
### Description

Sending messages to a topic doesn't work on iOS when the app is in
background. This branch will fix this by sending the messages to the
tokens of users subscribed to the topic, instead of sending them to the
topic directly. We can send a message to 500 devices per request, so in
most cases we won't need to make more than one request to firebase.

The PR add tests for notifications. Due to the usage of Firebase
messaging to send notifications, tests can not check notifications can
be sent. Instead, they make sure our endpoints to register/unregister a
device or subscribe to topics work.

To prevent anyone from subscribing to `bookingadmin` topic and then be
notified of all booking requests, the notification is now sent to all
members of the room manager group. We way want to send the notification
to just a few persons instead of the whole group.

Fix #291

---------

Co-authored-by: Armand Didierjean <95971503+armanddidierjean@users.noreply.github.com>