Revoke refresh token on password change (#708)
### Description

Currently, is someone account is compromised, the user can not log out
all active sessions.

Due to the use of a JWT we can not revoke active access token, but as
they have a limited duration they will automatically expire after some
time.

We should revoke refresh token to ensure that clients won't be able to
get a new access token.

As a consequence the user changing its password in settings will be
logged out of the app, but not immediately, only after the expiration of
its refresh token