Redirect to calypsso pages after confirmation (#722)
### Description

After an operation that involve opening an url, redirect to a nice
Success or Error page from CalypSSO instead of returning directly a
string.

For errors, this pull request only redirect for errors that are expected
to happen.
For example clicking an activation link a second times, or an expired
link should show a nice error. On the other hand we can expect that
opening a link with an nonexisting token should never happen under
normal conditions.