Move versions declaration in pyproject.toml (#700)
### Why

* Inspired by https://github.com/aeecleclair/Titan/pull/480
* Now both the version of itself and of the Titan client are declared in
the TOML config file

### How

* I tried cleverly adding
[`toml_file="pyproject.toml",`](https://docs.pydantic.dev/latest/concepts/pydantic_settings/#other-settings-source)
in the `SettingsConfigDict` (line 49) so that the model would load
fields from both config files (`.env` and `pyproject.toml`), but it
seems Pydantic does not allow that at all...
* So I went using Python's *native*
[`tomllib`](https://docs.python.org/3/library/tomllib.html) TOML parser
: it the only solution that worked for me, and it's short.
* (The `with` is out of the class because Pydantic started crying about
the type of the bytes flow from reading the file)

###
[`pyproject.toml`](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/)

* I just wanted to add
```toml
version = "4.3.3"
minimal-titan-version-code = 139
```
to the `[project]` table, but I realized it **didn't even exist**, so I
ended up writing the bare minimum.
* Perhaps we should fill this file seriously...
* it is powerful for automation and description (it could not hurt to
have it filled),
* and required to make a release on
[PyPI](https://pypi.org/user/ECLAIR/) (we already release CalypSSO and
the HA wrapper, why not Hyperion itself?).