Add trailing slash on `/users/account-types` (#825)
### Description

When visiting the visibility page on Titan, it loads indefinitely
because the HTTPS request `/users/account-types/` (notice the trailing
slash) hits a 307 to a HTTP request, given that Hyperion doesn't expect
it.

FastAPI redirects to HTTP when you issue a request with a trailing slash
that is unexpected or when you lack an expected trailing slash.
Although there are FastAPI-wide solutions (like an option in the
APIRouter or a flag in the CLI), this error doesn't seem to occur
anywhere else: Hyperion endpoints generally do NOT end with a trailing
slash, and Titan repositories generally use a suffix without trailing
slash, and when no suffix is used in a given Titan repo, the Hyperion
endpoint has this trailing slash.

Our front-end Titan has the trailing slash bescause the route has no
suffix (it's directly `final ext = "users/account-types/";`)

For evidence, see screenshot below of 2x2 requests (with and without
trailing slash, and with Hyperion instances that expect or not the
trailing slash):
- in prod: ` "/users/account-types"`
- in alpha: `"/users/account-types/"`

<img width="3119" height="730" alt="Capture d'écran 2025-09-06 113731"
src="https://github.com/user-attachments/assets/bdb98f57-a1d1-4fc1-9395-15aa9d69435b"
/>