Deploy/single stage docker (#852)
This pull request is part of a clarification process aimed at
simplifying handovers and maintainability.

Here, it is a clean-up of the Docker image build. 
The multi-stage build of the Docker image is not at all suitable in our
case, and was used to create a kind of ‘cache’, which can in fact be
easily resolved as mentioned in the FastAPI documentation as a Docker
‘trick’. The idea is to use --upgrade on pip install to take advantage
of layer caching. [Reference
FastAPI](https://fastapi.tiangolo.com/deployment/docker/?h=docke#docker-cache)
In addition, using a slim image means that only the strictly necessary
tools are included.

A multi-stage build image is relevant in the case of a compilation where
the code is subsequently executed in a lighter environment, which then
contains fewer vulnerabilities and is lighter.
The publish.yml has therefore been restored to its original version.

---------

Co-authored-by: Rayan <rbelahcene10@gmail.com>